feat(quantity): số lượng thập phân cho hàng cân ký (Đợt 4 d4-decimal: GL-07, POS-09) (#63)
* feat(quantity): số lượng thập phân cho hàng cân ký phía máy chủ (GL-07, POS-09)

- ADR-0015: cột số lượng và tồn kho numeric(14,3), cờ bán số lẻ trên mặt hàng và đơn vị quy đổi,
  một quy tắc tiền dòng round_half_up(đơn giá × số lượng) dùng chung máy chủ và PGlite.
- Migration 0062 đổi kiểu bằng USING col::numeric(14, 3), giữ dữ liệu.
- utils/quantity.ts: cộng trừ trên nghìn đơn vị, tiền × số lượng bằng BigInt; drizzle đọc numeric
  ở một chỗ (kiểu cột quantity()).
- Bán, trả, hủy đơn, phiếu nhập, trả hàng nhập, kiểm kê, điều chỉnh tồn, báo cáo đều nhận số lẻ và
  kiểm cờ; hàng không bật cờ gửi số lẻ trả 400.
- Nhập KiotViet tự bật cờ cho hàng đơn vị đo hoặc tồn lẻ; tồn đầu kỳ lẻ nhập được; tệp mẫu có cột
  Bán số lẻ.
- invariants.sql thêm I11: hàng không bật cờ thì tồn nguyên.


* feat(pos): bán số lẻ hàng cân ký trên POS, PGlite v006, in hóa đơn (GL-07, POS-09)

- PGlite v006: tồn, tồn biến thể, ngưỡng bậc giá sang NUMERIC(14,3), cờ allow_decimal_quantity
  cho sản phẩm và đơn vị quy đổi, xóa con trỏ để kéo lại danh mục. Máy chủ cũ chưa gửi cờ thì ghi false.
- Hàm isQuantityAllowed dùng chung máy chủ và POS.
- Ô QuantityInput nhận "1,5", hiển thị tối đa 3 chữ số lẻ; giỏ POS, phiếu nhập, trả hàng,
  trả hàng nhập, kiểm kê, bậc giá dùng chung. Tiền dòng qua lineAmount.
- Chi tiết phiếu nhập, dòng trả được, chi tiết phiếu kiểm trả kèm cờ số lẻ.
- formatQuantity ở hóa đơn in, chi tiết chứng từ, báo cáo, tồn kho.
- E2E bán 1,255 kg Nấm rơm ra 56.475đ rồi in hóa đơn A4; seed bật cờ cho Nấm rơm.


* fix(report): script TIEN-101 in số lượng numeric không kèm số 0 thừa (GL-07)

Cột số lượng đổi sang numeric(14,3) làm script in "3.000", dễ đọc nhầm thành ba nghìn.
Dùng trim_scale cho sl_mua và sl_da_tra.


* fix(pricing): ngưỡng chiết khấu danh mục nhận số lượng lẻ, không còn 500 (GL-07)

category_discounts.min_qty là integer nên Postgres suy tham số số lượng theo kiểu cột: bán 1,5 kg
cho khách có chiết khấu danh mục thì lỗi 500 ở tính giá, tạo đơn và đồng bộ ngoại tuyến.

- Đổi category_discounts.min_qty sang numeric(14,3) trong migration 0062 (USING, giữ dữ liệu);
  đầu vào zod vẫn nhận số nguyên.
- Ép ${quantity}::numeric ở câu so ngưỡng của chiết khấu danh mục và giá theo số lượng.
- PGlite v006 đổi catalog_category_discounts.min_qty sang NUMERIC(14,3), đọc bằng parseQuantity.
- Test: khách có chiết khấu 10%, bán 1,5 kg qua resolve-prices, POST /pos/orders và /sync/push.


* fix(product): biến thể đã xóa không chặn tắt bán số lẻ, I11 bỏ biến thể đã xóa (GL-07)

hasDecimalVariantStock đếm cả biến thể đã xóa nên tồn lẻ còn sót ở biến thể không bán được nữa
chặn tắt cờ. Thêm isNull(deletedAt), và I11 của invariants.sql lọc theo cùng điều kiện.


* fix(inventory): chặn 422 khi trả, hủy làm tồn thành số lẻ lúc cờ số lẻ đã tắt (GL-07)

Bán 1,5 kg rồi 0,5 kg (tồn 8), tắt cờ, trả 0,5 kg của đơn đầu thì tồn thành 8,5 ở mặt hàng không
bật cờ, vi phạm I11. Thêm assertStockStaysWhole: cờ tắt mà tồn mới của sản phẩm hoặc biến thể là
số lẻ thì 422 "Bật lại cho phép số lượng lẻ cho mặt hàng này để trả/hủy phần lẻ", rollback cả
chứng từ. Gọi ở trả hàng, hủy đơn, receiveStock và removeReceivedStock (hủy phiếu nhập, trả hàng
nhập).


* fix(quantity): ô số lượng đọc dấu chấm hàng nghìn theo vi-VN (GL-07)

parseQuantityInput coi mọi dấu chấm là thập phân nên "1.500" thành 1,5 (hoặc bị từ chối ở hàng
không bật cờ). Nay dấu phẩy là thập phân; dấu chấm theo sau đúng 3 chữ số, lặp nhóm 3, là phân cách
hàng nghìn ("1.500" là 1500, "1.250.000", "1.250,5"); dấu chấm khác vẫn là thập phân ("1.5", "0.255").


* fix(pricing): nhãn giá theo số lượng hiển thị "SL >= 2,5" (GL-07)


* docs(adr): ADR-0015 cập nhật ngưỡng chiết khấu danh mục, chặn tồn lẻ, dấu chấm hàng nghìn (GL-07)


* fix(quantity): ngưỡng giá, kiểm tồn giỏ và thông báo tồn của #62 tính số lẻ chính xác (GL-07)

Sau khi rebase lên POS-08/POS-13/POS-16 (#62) và POS-18 (#61):

- pricingBaseQuantity nhân hệ số quy đổi bằng mulQty: 0,58 bao 50 kg đúng bằng 29 kg, nhân float
  ra 28,999999999999996 thì trượt bậc giá và chiết khấu danh mục ngưỡng 29 (máy chủ và PGlite
  cùng lệch). Ngưỡng vẫn lọc trong resolvePriceFromSources theo quy tắc POS-16.
- stock-guard.ts cộng dồn tồn giỏ ở mọi đơn vị qua addQty/mulQty, use-add-to-cart quy đổi bằng
  mulQty: 0,1 kg cộng 0,1 túi 2 kg không còn bị chặn oan khi tồn đúng 0,3 kg. Thông báo vượt tồn
  in formatQuantity.
- Máy chủ: số tồn còn lại khi chặn đơn POS và vi phạm negative_stock_policy của đơn ngoại tuyến
  tính bằng addQty, in "0,3 kg" thay cho 0.30000000000000004.
- Test: bậc giá minQty 12 đạt khi bán 0,5 thùng 24, 0,58 bao 50 kg với bậc và chiết khấu danh
  mục 29, parity PGlite; giỏ số lẻ ở hai đơn vị; POS-13 số lẻ; bản sao v004 lên mới nhất chạy
  v005 rồi v006. Dòng sổ kho chèn tay trong test mang reference_type theo POS-18.
- ADR-0015: migration 0062, PGlite v006 sau v005, bất biến I11 (I10 là tham chiếu sổ kho của
  POS-18), quy tắc ngưỡng theo đơn vị cơ bản.


---------

+++ packages/shared/src/schema/product-history.ts
@@ -15,12 +15,12 @@ export const productPurchaseHistoryItemSchema = z.object({
   supplierName: z.string(),
   variantId: z.string().uuid().nullable(),
   variantLabelSnapshot: z.string().nullable(),
-  quantity: z.number().int(),
+  quantity: z.number(),
   unitPrice: z.number().int(),
   discountAmount: z.number().int(),
   lineTotal: z.number().int(),
   costAfter: z.number().int().nullable(),
-  stockAfter: z.number().int().nullable(),
+  stockAfter: z.number().nullable(),
 })
 
 export const productStockCheckHistoryItemSchema = z.object({
@@ -32,9 +32,9 @@ export const productStockCheckHistoryItemSchema = z.object({
   adjustedByName: z.string().nullable(),
   variantId: z.string().uuid().nullable(),
   variantLabelSnapshot: z.string().nullable(),
-  systemQty: z.number().int(),
-  actualQty: z.number().int(),
-  diff: z.number().int(),
+  systemQty: z.number(),
+  actualQty: z.number(),
+  diff: z.number(),
   note: z.string().nullable(),
 })
